Move helper methods to view_data.py
