+++ codespeed/views_data.py
@@ -3,6 +3,8 @@ from __future__ import absolute_import
 
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
+from django.shortcuts import get_object_or_404
+
 from codespeed.models import (
     Executable, Revision, Project, Branch,
     Environment, Benchmark, Result)
@@ -229,3 +231,32 @@ def get_benchmark_results(data):
             'results': result_list,
             'relative': relative_results,
            }
+
+
+def get_num_revs_and_benchmarks(data):
+    if data['ben'] == 'grid':
+        benchmarks = Benchmark.objects.all().order_by('name')
+        number_of_revs = 15
+    elif data['ben'] == 'show_none':
+        benchmarks = []
+        number_of_revs = int(data.get('revs', 10))
+    else:
+        benchmarks = [get_object_or_404(Benchmark, name=data['ben'])]
+        number_of_revs = int(data.get('revs', 10))
+    return number_of_revs, benchmarks
+
+
+def get_stats_with_defaults(res):
+    val_min = ""
+    if res.val_min is not None:
+        val_min = res.val_min
+    val_max = ""
+    if res.val_max is not None:
+        val_max = res.val_max
+    q1 = ""
+    if res.q1 is not None:
+        q1 = res.q1
+    q3 = ""
+    if res.q3 is not None:
+        q3 = res.q3
+    return q1, q3, val_max, val_min
